Add files via upload
The RSSI dataset has 13 beacon readings which tell the strength of the signal, RSSI readings are represented in negative,  if it is –50 it represents the user is close to the ibeacon and hence the strength of the signal is much strong than the value –95 which represents the user isn't in the close proximity of the beacon and hence the strength of the signal is weak.  BLE RSSI labeled dataset is used to train different classifiers and analyze the performance.